requirements.txt
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.5
numpy==1.26.4
opencv-python
Pillow
torch>=1.7
torchvision
tqdm
timm
pyiqa==0.1.12
cupy-cuda12x==13.3.0
pytorch_msssim